# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WhatsApp API Clone",
        "description": "This is the API documentation for the WhatsApp clone application.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/messages/send": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "Send a message to a chatroom",
                "operationId": "8a5fa086537e5c967d8a93bb0453cf0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "chatroom_id"
                                ],
                                "properties": {
                                    "chatroom_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message_text": {
                                        "type": "string",
                                        "example": "Hello, World!"
                                    },
                                    "attachment": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "id",
                                        "chatroom_id",
                                        "user_id",
                                        "message_text"
                                    ],
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "chatroom_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message_text": {
                                            "type": "string",
                                            "example": "Hello, World!"
                                        },
                                        "attachment_path": {
                                            "type": "string",
                                            "example": "http://example.com/file.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Either message text or an attachment is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Something went wrong. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/messages/{id}": {
            "get": {
                "tags": [
                    "Message"
                ],
                "summary": "Get messages from a chatroom",
                "operationId": "dc59b3d4bdbd0bc343251aa9acb502e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Chatroom ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of messages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "required": [
                                            "id",
                                            "chatroom_id",
                                            "user_id",
                                            "message_text"
                                        ],
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "chatroom_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "message_text": {
                                                "type": "string",
                                                "example": "Hello, World!"
                                            },
                                            "attachment_path": {
                                                "type": "string",
                                                "example": "http://example.com/file.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chatroom not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Chatroom not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}